4.15 - 4.22 exercises, added jsonwebtoken, bcrypt, user model, middlewares and new routes

+++ part4/blog-list/controllers/blog-ctrl.js
@@ -1,5 +1,7 @@
 require('express-async-errors')
+const jwt = require('jsonwebtoken')
 const Blog = require('../models/Blog.js')
+const User = require('../models/User.js')
 
 const getAll = async (req, res, next) => {
   const data = await Blog.find({})
@@ -13,14 +15,27 @@ const create = async (req, res) => {
     })
   }
 
+  const decodedToken = jwt.verify(req.token, process.env.SECRET)
+
+  if (!req.token || !decodedToken.id) {
+    return res.status(401).json({
+      message: 'token missing or invalid'
+    })
+  }
+
+  const user = await User.findById(decodedToken.id)
+
   const blog = new Blog({
     title: req.body.title,
     author: req.body.author,
     url: req.body.url,
-    likes: req.body.likes
+    likes: req.body.likes,
+    user: user._id
   })
 
   const data = await blog.save()
+  user.blogs = user.blogs.concat(data._id)
+  await user.save()
 
   res.status(201).json(data)
 }
@@ -28,8 +43,22 @@ const create = async (req, res) => {
 const deleteById = async (req, res) => {
   const id = req.params.id
 
-  const data = await Blog.findByIdAndRemove(id)
-  res.status(204).json(data)
+  const decodedToken = jwt.verify(req.token, process.env.SECRET)
+
+  if (!req.token || !decodedToken.id) {
+    return res.status(401).json({
+      message: 'token missing or invalid'
+    })
+  }
+
+  const blog = await Blog.findById(id)
+
+  if (blog.user.toString() === decodedToken.id.toString()) {
+    await Blog.findByIdAndRemove(id)
+    res.status(204).end()
+  } else {
+    res.status(401).json({ message: 'you don\'t have permissions to do that' })
+  }
 }
 
 const update = async (req, res) => {
